Adding the ability to snapshot multiple accounts within an organization.

+++ aws_data_getter.py
@@ -1,4 +1,5 @@
-from awshelper import aws_get_config, aws_get_regions, aws_init, aws_test_access
+import awshelper
+import boto3
 from argparse import ArgumentParser, ArgumentTypeError
 import json
 import os
@@ -20,12 +21,12 @@ def _str2bool(v):
         raise ArgumentTypeError('Boolean value expected.')
 
 
-def snapshot_configs(output_folder):
-    regions = aws_get_regions()  # empty list if we are not doing aws
+def snapshot_configs(output_folder, account=''):
+    regions = awshelper.aws_get_regions()  # empty list if we are not doing aws
     for region in regions:
-        region_dir = os.path.join(output_folder, AWS_SUB_FOLDER, region)
+        region_dir = os.path.join(output_folder, AWS_SUB_FOLDER, account, region)
         os.makedirs(region_dir)
-        aws_region_config = aws_get_config(region)
+        aws_region_config = awshelper.aws_get_config(region)
         for key in aws_region_config:
             try:
                 filename = os.path.join(region_dir, key + ".json")
@@ -39,7 +40,7 @@ def main():
     parser = ArgumentParser(description="archiver")
     parser.add_argument('-c', '--config', dest='config_filename', help="config file", metavar="<file>",
                         default="config.json")
-    parser.add_argument('-p', '--profile', dest='profile', help="profile name")
+    parser.add_argument('-p', '--profile', dest='profile', help="profile name", default=None)
     parser.add_argument('-o', '--output-folder', dest='output_folder', help="output folder", metavar="<file>",
                         default="aws-snapshot")
     parser.add_argument('-t', '--test-access', dest='test_access', help="Test access to AWS and exit", type=_str2bool,
@@ -54,11 +55,14 @@ def main():
     regions = options.get("regions", None)
     vpc_ids = options.get("vpcs", None)
     skip_data = options.get("skipData", None)
-
+    accounts = options.get("accounts", None)
+    defaultRole = options.get("defaultRole", None)
+    useAccountName = options.get("useAccountName", False)
     if args.test_access:
         aws_test_access(profile=args.profile)
         exit(0)
-
+    if args.profile and accounts != None:
+        raise Exception("Cannot use option -p along with config option accounts")
     if os.path.exists(args.output_folder):
         if args.force:
             shutil.rmtree(args.output_folder)
@@ -67,10 +71,24 @@ def main():
                 "Output folder '{}' already exists. Provide a different folder (using '-o' option) or use the '-f' option to overwrite existing folder".format(
                     args.output_folder), file=sys.stderr)
             exit(1)
-
-    aws_init(regions, vpc_ids, skip_data, args.profile)
-    snapshot_configs(args.output_folder)
-
+    sessions = []
+    if accounts != None:  
+        for account in awshelper.get_aws_accounts(accounts):
+            role = next((entry.get("role", defaultRole) for entry in accounts if entry["id"] == account['Id']),defaultRole)
+            session = awshelper.get_aws_sessions(account['Id'], role)
+            if session:
+                if useAccountName:
+                    name = account['Name']
+                else:
+                    name = account['Id']
+                sessions.append((name, session))
+    else:
+        session = boto3.session.Session(profile_name=args.profile)
+        sessions.append((session.profile_name, session))
+    for session in sessions:
+        print(f"Processing account: {session[0]}")
+        awshelper.aws_init(regions, vpc_ids, skip_data, session[1])
+        awshelper.snapshot_configs(args.output_folder, session[0])
 
 if __name__ == '__main__':
     main()
